move hooks; remove debug code
